Implement test_google_ecommerce_purchase
